refactor: relocation sequelize, passport to utils
passport 와 sequelize 가 services 레이어에 있는 것이 어색하다고 판단하여
진혁님과 이야기 한 후 utils 로 옮김
