+++ server/bin/www.ts
@@ -1,6 +1,6 @@
 import '../src/env';
 import app from '../src/app';
-import '../src/services/sequelize';
+import '../src/utils/sequelize';
 
 (() => {
   app.listen(3000);
