+++ server/src/app.ts
@@ -2,7 +2,7 @@ import * as express from 'express';
 import * as cookieParser from 'cookie-parser';
 import * as passport from 'passport';
 
-import setUpPassport from '../src/services/passport';
+import setUpPassport from './utils/passport';
 import indexRouter from './routes/api';
 
 const app = express();
